# Restored the GMF data structure used in engine 2.0