--- a/openquake/calculators/event_based.py
+++ b/openquake/calculators/event_based.py
@@ -464,8 +464,7 @@ class EventBasedRuptureCalculator(PSHACalculator):
 
 # ######################## GMF calculator ############################ #
 
-gmv_dt = numpy.dtype(
-    [('gmv', F32), ('eid', U32), ('rlzi', U16), ('imti', U8)])
+gmv_dt = numpy.dtype([('sid', U32), ('eid', U32), ('imti', U8), ('gmv', F32)])
 
 
 def compute_gmfs_and_curves(getter, rlzs, monitor):
@@ -485,18 +484,19 @@ def compute_gmfs_and_curves(getter, rlzs, monitor):
    """
     oq = monitor.oqparam
     haz = {sid: {} for sid in getter.sids}
-    gmfcoll = {sid: [] for sid in getter.sids}
+    gmfcoll = {}  # rlz -> gmfa
     for rlz in rlzs:
+        gmfcoll[rlz] = []
         for sid, gmvdict in zip(getter.sids, getter(rlz)):
             if gmvdict:
                 for imti, imt in enumerate(getter.imts):
                     if oq.hazard_curves_from_gmfs:
                         haz[sid][imt, rlz] = gmvdict[imt]
                     for rec in gmvdict[imt]:
-                        gmfcoll[sid].append(
-                            (rec['gmv'], rec['eid'], rlz.ordinal, imti))
-    for sid in gmfcoll:
-        gmfcoll[sid] = numpy.array(gmfcoll[sid], gmv_dt)
+                        gmfcoll[rlz].append(
+                            (sid, rec['eid'], imti, rec['gmv']))
+    for rlz in gmfcoll:
+        gmfcoll[rlz] = numpy.array(gmfcoll[rlz], gmv_dt)
     result = dict(gmfcoll=gmfcoll if oq.ground_motion_fields else None,
                   hcurves={})
     if oq.hazard_curves_from_gmfs:
@@ -537,9 +537,10 @@ class EventBasedCalculator(ClassicalCalculator):
         agg_mon = self.monitor('aggregating hcurves')
         if res['gmfcoll'] is not None:
             with sav_mon:
-                for sid, array in res['gmfcoll'].items():
+                for rlz, array in res['gmfcoll'].items():
                     if len(array):
-                        self.datastore.extend('gmf_data/sid-%04d' % sid, array)
+                        key = 'gmf_data/%04d' % rlz.ordinal
+                        self.datastore.extend(key, array)
         slicedic = self.oqparam.imtls.slicedic
         with agg_mon:
             for key, poes in res['hcurves'].items():