--- a/openquake/calculators/tests/event_based_test.py
+++ b/openquake/calculators/tests/event_based_test.py
@@ -24,6 +24,7 @@ from nose.plugins.attrib import attr
 
 import numpy.testing
 
+from openquake.baselib.general import group_array
 from openquake.commonlib.datastore import read
 from openquake.commonlib.util import max_rel_diff_index
 from openquake.commonlib.export import export
@@ -89,8 +90,9 @@ class EventBasedTestCase(CalculatorTestCase):
             oq = self.calc.oqparam
             self.assertEqual(list(oq.imtls), ['PGA'])
             dstore = read(self.calc.datastore.calc_id)
-            gmvs_site_0 = dstore['gmf_data/sid-0000']['gmv']
-            gmvs_site_1 = dstore['gmf_data/sid-0001']['gmv']
+            gmf = group_array(dstore['gmf_data/0000'], 'sid')
+            gmvs_site_0 = gmf[0]['gmv']
+            gmvs_site_1 = gmf[1]['gmv']
             joint_prob_0_5 = joint_prob_of_occurrence(
                 gmvs_site_0, gmvs_site_1, 0.5, oq.investigation_time,
                 oq.ses_per_logic_tree_path)