In master I changed the way GMFs are stored, to be efficient in the case "multiple realizations, single site". However we later changed the risk calculator to work "one realization, multiple sites", i.e. in the opposite regime. This is why the GMF data structure has to be changed back the way it was. Here is the difference in the case of Miriam's island on the cluster:
```
   master: total compute_gmfs_and_curves=536 s, saving gmfs=126s
   this branch: total compute_gmfs_and_curves=699s, saving_gmfs=0.439s
```
Since `total compute_gmfs_and_curves` runs in parallel, it does not matter if it is a only 22% slower; on the other hand `saving_gmfs` runs on a single core and having it 287 times faster is a game changer.